Chapter "Context"
This was indeed the first chapter that was really confusing to me. Maybe
it's because I didn't understand exactly what this HTTPHandler does yet,
but this `context` thing is indeed very confusing and I have never seem
this in another programming language. I'll definitely have to search for
extra material about this.