Chapter "Maths"
This chapter was the most boring so far. I don't know... It seems like
it didn't really show any new features from the language itself and
neither dived deep into the math library or any particularity of it that
distinguishes it from math libraries of other languages. Also, the
problem of drawing a clock face was so overcomplicated and repeatedly
that I just jumped some parts of the chapter and went to the parts that
seemed relevant.